CIV-7263 Stop swallowing exception silently
A mistake was made when copy/pasting this code from prl-cos-api to civil-general-application and that mistake ended up flowing to civil-service as a result.
The mistake is simply the removal of a rethrow clause, whose absence causes errors to be swallowed silently and not being reported to PayHUB.
This fixes the issue by re-introducing the missing throw clause, so that we will report HTTP-500 exceptions to PayHUB which will in turn attempt to notify us again.