https://tools.hmcts.net/jira/browse/DTSCCI-2514
Merge the relevant packages and functionality from the civil-commons repository directly into the civil-service repository.

update 1: checkstyle fix: The classes not passing the check in civil-commons originally used .* imports instead of direct class imports in 11 test files.

update 2: checkstyle fix: The classes not passing the check in civil-commons originally missed the @author and @version from javadoc and removed the https://www. part from the link as it shouldn't be specified as [plain text